+++ test/date_time/date_range_picker_test.dart
@@ -46,4 +46,62 @@ void main() {
     final dateRangePickerText = dateRangPickerTextBuilder(state);
     expect(find.text(dateRangePickerText.data!), findsOneWidget);
   });
+
+  testWidgets('shows input date range picker dialog on GestureDetector tap',
+      (tester) async {
+    await tester.pumpWidget(buildMaterialTestApp([
+      FastDateRangePicker(
+        name: 'date_range_picker',
+        firstDate: DateTime(1900),
+        lastDate: DateTime.now().add(const Duration(days: 365)),
+      ),
+    ]));
+
+    await tester.tap(findInkWell().first);
+    await tester.pumpAndSettle();
+
+    expect(find.byType(DateRangePickerDialog), findsOneWidget);
+  });
+
+  testWidgets('shows calendar date range picker dialog on calendar button tap',
+      (tester) async {
+    await tester.pumpWidget(buildMaterialTestApp([
+      FastDateRangePicker(
+        name: 'date_range_picker',
+        firstDate: DateTime(1900),
+        lastDate: DateTime.now().add(const Duration(days: 365)),
+      ),
+    ]));
+
+    final calenderButton = find.widgetWithIcon(IconButton, Icons.today);
+    await tester.tap(calenderButton);
+    await tester.pumpAndSettle();
+
+    expect(find.byType(DateRangePickerDialog), findsOneWidget);
+  });
+
+  testWidgets('tracks touched state', (tester) async {
+    await tester.pumpWidget(buildMaterialTestApp([
+      FastDateRangePicker(
+        name: 'date_range_picker',
+        firstDate: DateTime(1900),
+        lastDate: DateTime.now().add(const Duration(days: 365)),
+      ),
+    ]));
+
+    final state =
+        tester.state<FastDateRangePickerState>(findFastDateRangePicker());
+
+    final calenderButton = find.widgetWithIcon(IconButton, Icons.today);
+    await tester.tap(calenderButton);
+    await tester.pumpAndSettle();
+
+    expect(state.touched, false);
+
+    final closeButton = find.widgetWithIcon(IconButton, Icons.close);
+    await tester.tap(closeButton);
+    await tester.pumpAndSettle();
+
+    expect(state.touched, true);
+  });
 }